chore(dashboards): move the metric card template tiles into the dashboards product
