+++ frontend/src/scenes/dashboard/newDashboardLogic.ts
@@ -39,10 +39,11 @@ import {
     JsonType,
 } from '~/types'
 
+import { WEBSITE_METRICS_METRIC_CARD_TILES } from 'products/dashboards/frontend/websiteMetricsMetricCardTemplate'
+
 import type { FeatureFlagsSet } from '../../lib/logic/featureFlagLogic'
 import type { InsightModel } from '../../types'
 import { UNFILED_DASHBOARDS_FOLDER } from './dashboardConstants'
-import { WEBSITE_METRICS_METRIC_CARD_TILES } from './websiteMetricsMetricCardTemplate'
 
 export interface NewDashboardForm {
     name: string
